Flytter utkast-kall i egen fil
